midb: more sensible variable types for summary_folder/P-FDDT

+++ mra/imap/imap.hpp
@@ -198,7 +198,7 @@ extern BOOL (*system_services_add_user_into_temp_list)(const char *, int);
 extern authmgr_login_t system_services_auth_login;
 extern gromox::errno_t (*system_services_auth_meta)(const char *username, unsigned int wantpriv, sql_meta_result &out);
 extern int (*system_services_get_uid)(const char *, const char *, const std::string &, unsigned int *);
-extern int (*system_services_summary_folder)(const char *, const char *, int *, int *, int *, unsigned long*, unsigned int *, int *);
+extern int (*system_services_summary_folder)(const char *, const char *, size_t *, size_t *, size_t *, uint32_t *, uint32_t *, int *);
 extern int (*system_services_make_folder)(const char *, const char *, int *);
 extern int (*system_services_remove_folder)(const char *, const char *, int *);
 extern int (*system_services_rename_folder)(const char *, const char *, const char *, int *);
